refactor: создал файл для заглушек и функций помощников
