+++ tape_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func TestTape_Write(t *testing.T) {
-	file, clean := createTempFile(t, "12345")
+	file, clean := CreateTempFile(t, "12345")
 	defer clean()
 
 	tape := &tape{file}
